Refactor page.tsx into focused hooks and components
Extract 5 custom hooks, 1 component, and 1 utility from the 1600-line
page.tsx to improve maintainability:

- hooks/useThinkingState.ts — thinking indicator state and animations
- hooks/useScrollManager.ts — scroll tracking, auto-scroll, morph bar
- hooks/usePullToRefresh.ts — pull-to-refresh gesture handling
- hooks/useKeyboardLayout.ts — iOS Safari keyboard offset
- hooks/useTheme.ts — light/dark theme with localStorage
- components/ChatHeader.tsx — header bar with status indicators
- lib/parseBackendModels.ts — model parsing utility

page.tsx reduced from ~1600 to ~1050 lines. All 64 tests pass, build
succeeds with no changes to runtime behavior.
