+++ hooks/useTheme.ts
@@ -1,0 +1,27 @@
+import { useState, useEffect, useCallback } from "react";
+
+/**
+ * Manages light/dark theme state with localStorage persistence
+ * and document class synchronization.
+ */
+export function useTheme() {
+  const [theme, setTheme] = useState<"light" | "dark">("light");
+
+  useEffect(() => {
+    const isDark = document.documentElement.classList.contains("dark");
+    setTheme(isDark ? "dark" : "light");
+  }, []);
+
+  const toggleTheme = useCallback(() => {
+    setTheme((prev) => {
+      const next = prev === "light" ? "dark" : "light";
+      document.documentElement.classList.toggle("dark", next === "dark");
+      try {
+        localStorage.setItem("theme", next);
+      } catch {}
+      return next;
+    });
+  }, []);
+
+  return { theme, toggleTheme };
+}
